@@ -7,7 +7,7 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "ilpy"
 description = "Python wrappers for popular MIP solvers."
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 license = { text = "MIT" }
 authors = [
     { email = "funkej@janelia.hhmi.org", name = "Jan Funke" },
@@ -48,7 +48,7 @@ readme = { file = ["README.md"] }
 
 # https://beta.ruff.rs/docs
 [tool.ruff]
-target-version = "py38"
+target-version = "py39"
 src = ["ilpy"]
 
 [tool.ruff.lint]
